Render @-mentions in command chat messages and explicitly mention all context
For commands (like Explain Code and Find Code Smells), the prompt is displayed poorly in the chat UI:

- @-mentions are not rendered
- Implicit context is included (the full file) that is not shown in the message, so if you resend the message, it resends without that implicit context.

This change fixes that, showing @-mentions and including the other context mentioned.

To accomplish this, the extension code (not running in the webview) needs to be able to parse the Lexical editor state, so much of that code moves to `lib/shared`.

Fixes https://linear.app/sourcegraph/issue/CODY-1984/commands-do-not-render-mentions-and-include-implicit-context